requirements.txt
google-api-python-client>=2.0
selenium>=4.0
python-dotenv>=1.0
